Antenna Builder .mat file and .m file export
Dumbbell.mat is the data file from the antennaBuilder MATLAB toolbox. 

seven_mhz_dumbbell.m is the exported file from the antennaBuilder that includes the frequency band information and allows us to show the visualizations of a wire antenna version of the design with some hand-edited-in commands. 

If you want to experiment directly in antennaBuilder, then import the .mat file. If you want to produce a wire antenna object version of the design, then use the .m file.