updated model
Action Item for next time - Need to change the name since the design isn't for a particular frequency. 

This model has several improvements. The math is cleared up and there are a lot of additional comments explaining the different parts. There are more visualizations and the broken visualizations were removed. There's an animation of the electric fields at the end of the model. Sections were added to make it easier to debug.